Reorganizes source and test layout.
